Fix: Changes in file structute to follow repository design patterns
