+++ package.json
@@ -34,6 +34,7 @@
     "@types/bcrypt": "^3.0.0",
     "@types/express": "^4.17.8",
     "@types/mongoose": "^5.7.36",
+    "@types/node": "^14.11.2",
     "@types/passport": "^1.0.4",
     "@types/passport-jwt": "^3.0.3",
     "@types/passport-local": "^1.0.33",
